requirements.txt
flask>=3.0.0
tensorflow>=2.15.0
pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.8.0
werkzeug>=3.0.0
